+++ Microsoft.Azure.Cosmos/src/ChangeFeedProcessor/FeedProcessing/FeedProcessorCore.cs
@@ -8,6 +8,7 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.FeedProcessing
     using System.Collections.Generic;
     using System.Collections.ObjectModel;
     using System.Diagnostics;
+    using System.Linq;
     using System.Net;
     using System.Threading;
     using System.Threading.Tasks;
@@ -24,20 +25,20 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.FeedProcessing
         private readonly PartitionCheckpointer checkpointer;
         private readonly ChangeFeedObserver<T> observer;
         private readonly FeedIterator resultSetIterator;
-        private readonly CosmosSerializer cosmosJsonSerializer;
+        private readonly CosmosSerializerCore serializerCore;
 
         public FeedProcessorCore(
             ChangeFeedObserver<T> observer,
             FeedIterator resultSetIterator,
             ProcessorOptions options,
             PartitionCheckpointer checkpointer,
-            CosmosSerializer cosmosJsonSerializer)
+            CosmosSerializerCore serializerCore)
         {
             this.observer = observer;
             this.options = options;
             this.checkpointer = checkpointer;
             this.resultSetIterator = resultSetIterator;
-            this.cosmosJsonSerializer = cosmosJsonSerializer;
+            this.serializerCore = serializerCore;
         }
 
         public override async Task RunAsync(CancellationToken cancellationToken)
@@ -114,10 +115,12 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.FeedProcessing
         private Task DispatchChangesAsync(ResponseMessage response, CancellationToken cancellationToken)
         {
             ChangeFeedObserverContext context = new ChangeFeedObserverContextCore<T>(this.options.LeaseToken, response, this.checkpointer);
-            Collection<T> asFeedResponse;
+            IEnumerable<T> asFeedResponse;
             try
             {
-                asFeedResponse = this.cosmosJsonSerializer.FromStream<CosmosFeedResponseUtil<T>>(response.Content).Data;
+                asFeedResponse = this.serializerCore.FromFeedResponseStream<T>(
+                    response.Content,
+                    Documents.ResourceType.Document);
             }
             catch (Exception serializationException)
             {
@@ -126,13 +129,12 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.FeedProcessing
             }
 
             // When StartFromBeginning is used, the first request returns OK but no content
-            if (asFeedResponse.Count == 0)
+            if (!asFeedResponse.Any())
             {
                 return Task.CompletedTask;
             }
 
-            List<T> asReadOnlyList = new List<T>(asFeedResponse.Count);
-            asReadOnlyList.AddRange(asFeedResponse);
+            List<T> asReadOnlyList = new List<T>(asFeedResponse);
 
             return this.observer.ProcessChangesAsync(context, asReadOnlyList, cancellationToken);
         }